pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name="domainator"
authors = [
    {name = "Sean Johnson", email = "sjohnson@neb.com"},
    {name = "Andrew Ge"},
    {name = "Yu-Cheng Lin"},
    {name = "Zhiyi Sun"}
]
description="A flexible and modular software suite for domain-based gene neighborhood and protein search, extraction, and clustering."
readme = "README.md"
requires-python = ">=3.9"

classifiers=[
    "Programming Language :: Python :: 3",
]
license = {file = "LICENSE.txt"}
dynamic = ["version"]
dependencies = [
    "pandas >=2.1.0",
    "seaborn >=0.13.0",
    "scipy >=1.11.2",
    "pyhmmer >=0.10.2 ",
    "umap-learn >=0.5.4",
    "h5py >=3.9.0",
    "jsonargparse >=4.18.0",
    "psutil >=5.9.6",
    "tqdm >=4.65.0",
    "pyrodigal >=3.0.1",
    "bashplotlib >=0.6.5",
    "requests >=2.31.0"
]

[project.optional-dependencies]
test = ["pytest >=7.4.2", 
        "pytest-datadir~=1.4.1"]
# add dependency for Foldseek/ESM2 related stuff.

[project.urls]
Bug_Tracker = "https://github.com/seanrjohnson/domainator/issues"


[tool.setuptools]
package-dir = {"" = "src"}


[tool.setuptools.dynamic]
version = {attr = "domainator.__version__"}

[project.scripts]
"build_projection.py" = "domainator:build_projection._entrypoint"
"build_ssn.py" = "domainator:build_ssn._entrypoint"
"build_tree.py" = "domainator:build_tree._entrypoint"
"color_genbank.py" = "domainator:color_genbank._entrypoint"
"color_table_to_legend.py" = "domainator:color_table_to_legend._entrypoint"
"compare_contigs.py" = "domainator:compare_contigs._entrypoint"
"deduplicate_genbank.py" = "domainator:deduplicate_genbank._entrypoint"
"domain_search.py" = "domainator:domain_search._entrypoint"
"domainate.py" = "domainator:domainate._entrypoint"
"domainator_db_download.py" = "domainator:domainator_db_download._entrypoint"
"enum_report.py" = "domainator:enum_report._entrypoint"
"extract_domains.py" = "domainator:extract_domains._entrypoint"
"extract_peptides.py" = "domainator:extract_peptides._entrypoint"
"extract_unannotated.py" = "domainator:extract_unannotated._entrypoint"
"filter_domains.py" = "domainator:filter_domains._entrypoint"
"genbank_to_fasta.py" = "domainator:genbank_to_fasta._entrypoint"
"hmmer_build.py" = "domainator:hmmer_build._entrypoint"
"hmmer_compare.py" = "domainator:hmmer_compare._entrypoint"
"hmmer_report.py" = "domainator:hmmer_report._entrypoint"
"hmmer_search.py" = "domainator:hmmer_search._entrypoint"
"hmmer_select.py" = "domainator:hmmer_select._entrypoint"
"matrix_report.py" = "domainator:matrix_report._entrypoint"
"partition_seqfile.py" = "domainator:partition_seqfile._entrypoint"
"select_by_cds.py" = "domainator:select_by_cds._entrypoint"
"select_by_contig.py" = "domainator:select_by_contig._entrypoint"
"seq_dist.py" = "domainator:seq_dist._entrypoint"
"summary_report.py" = "domainator:summary_report._entrypoint"
"transform_matrix.py" = "domainator:transform_matrix._entrypoint"
"plot_contigs.py" = "domainator:plot_contigs._entrypoint"
"trim_contigs.py" = "domainator:trim_contigs._entrypoint"